nexus: move Describe map-clone test into operation_test.go
Consolidate the test into the package's existing operation_test.go
rather than a separate file, matching the convention used for the
scheduler fix.
